+++ packages/dash-html-components/scripts/generate-components.js
@@ -57,14 +57,14 @@ function generatePropTypes(element, attributes) {
      * An integer that represents the number of times
      * that this element has been clicked on.
      */
-    'n_clicks': PropTypes.integer,
+    'n_clicks': PropTypes.number,
 
     /**
      * An integer that represents the time (in ms since 1970)
      * at which n_clicks changed. This can be used to tell
      * which button was changed most recently.
      */
-    'n_clicks_timestamp': PropTypes.integer,
+    'n_clicks_timestamp': PropTypes.number,
 
     /**
      * A unique identifier for the component, used to improve
@@ -120,6 +120,7 @@ function generateComponent(Component, element, attributes) {
     return `
 import React from 'react';
 import PropTypes from 'prop-types';
+import {omit} from 'ramda';
 
 const ${Component} = (props) => {
     return (
@@ -133,7 +134,7 @@ const ${Component} = (props) => {
                 }
                 if (props.fireEvent) props.fireEvent({event: 'click'});
             }}
-            {...props}
+            {...omit(['n_clicks', 'n_clicks_timestamp'], props)}
         >
             {props.children}
         </${element}>