+++ packages/dash-html-components/src/components/Base.react.js
@@ -1,6 +1,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import {omit} from 'ramda';
 
 const Base = (props) => {
     return (
@@ -14,7 +15,7 @@ const Base = (props) => {
                 }
                 if (props.fireEvent) props.fireEvent({event: 'click'});
             }}
-            {...props}
+            {...omit(['n_clicks', 'n_clicks_timestamp'], props)}
         >
             {props.children}
         </base>
@@ -43,14 +44,14 @@ Base.propTypes = {
      * An integer that represents the number of times
      * that this element has been clicked on.
      */
-    'n_clicks': PropTypes.integer,
+    'n_clicks': PropTypes.number,
 
     /**
      * An integer that represents the time (in ms since 1970)
      * at which n_clicks changed. This can be used to tell
      * which button was changed most recently.
      */
-    'n_clicks_timestamp': PropTypes.integer,
+    'n_clicks_timestamp': PropTypes.number,
 
     /**
      * A unique identifier for the component, used to improve